% Generated manually
\name{color_dis_default}
\alias{color_dis_default}
\docType{data}
\title{Default Discrete Color Palette}
\description{
A vector of 11 color hex codes used as the default discrete color scale
throughout the \pkg{GeneLowPaper} package.
}
\usage{color_dis_default}
\format{A character vector of length 11.}
\examples{
 color_dis_default
}
\author{GeneLowPaper package}
