% Generated by generate_rd.py
\name{gene_Cellchat2}
\alias{gene_Cellchat2}
\title{Single-Gene Differential Cell Communication (Ligand-Receptor)
}
\description{
Compares cell-cell communication between gene-high and gene-low groups

within a target cell type. Outputs comparison bar plots, scatter plots of

signaling roles, and ranked signaling pathway differences.
}
\usage{
gene_Cellchat2(scPath = NULL, gene = "TP53", Tar_celltype = "Malignant", outdir = "./", plotWidth = 8, plotHeight = 6, color_dis = color_dis_default)
}
\arguments{
  \item{scPath}{Character. Path to the pre-processed Seurat object RDS file.}
  \item{gene}{Character. Target gene symbol.}
  \item{Tar_celltype}{Character. Cell type to split. Default is \code{"Malignant"}.}
  \item{outdir}{Character. Output directory. Default is \code{"./"}.}
  \item{plotWidth}{Numeric. Plot width. Default is 8.}
  \item{plotHeight}{Numeric. Plot height. Default is 6.}
  \item{color_dis}{Character vector. Color palette. Defaults to \code{color_dis_default}.}
}
\details{
Output directory: \code{outdir/6_Fig/cellchat/}.

Temporary RDS files \code{tmp_Low.Rds} and \code{tmp_High.Rds} are created

in the working directory.


}
\value{
None. Saves comparison plots to disk.
}
\examples{
\dontrun{

gene_Cellchat2(scPath = "NSCLC_GSE148071_Single.Rds", gene = "IL36RN",

Tar_celltype = "Malignant")

}

}
\author{GeneLowPaper package}
