% Generated by generate_rd.py
\name{sc_dorotheaTF}
\alias{sc_dorotheaTF}
\title{Dorothea Transcription Factor Activity Analysis
}
\description{
Computes transcription factor activities using dorothea regulons and decoupleR

for a target gene-defined cell group, then visualizes top differential TFs.
}
\usage{
sc_dorotheaTF(scPath = NULL, gene = "TP53", Tar_celltype = "Malignant", topTF = 25, outdir = "./", plotWidth = 8, plotHeight = 4, color_dis = color_dis_default)
}
\arguments{
  \item{scPath}{Character. Path to the pre-processed Seurat object RDS file.}
  \item{gene}{Character. Target gene symbol.}
  \item{Tar_celltype}{Character. Cell type to subset and analyze. Default is \code{"Malignant"}.}
  \item{topTF}{Numeric. Number of top TFs to display. Default is 25.}
  \item{outdir}{Character. Output directory. Default is \code{"./"}.}
  \item{plotWidth}{Numeric. Plot width. Default is 8.}
  \item{plotHeight}{Numeric. Plot height. Default is 4.}
  \item{color_dis}{Character vector. Color palette. Defaults to \code{color_dis_default}.}
}
\details{
Output directory: \code{outdir/6_Fig/Gene_TF/}.


}
\value{
None. Saves TF heatmap and result CSV to disk.
}
\author{GeneLowPaper package}
